fix(runs,tui,cli): relocate the run-inventory reader to core so ls works without the [tui] extra (#650)
`bmad-loop ls` on a core-only install crashed with ModuleNotFoundError: no
module named 'pyte'. cmd_list lazily imported tui.data.discover_runs, and
data.py imports pyte + rich at module scope; the inline comment guarding that
import ("data.py has no textual imports") was literally true and named the
wrong dependency. Broken since v0.6.1.

Move discover_runs and its whole closure -- the status constants, _stat_sig,
liveness, _session_liveness, _classify, RunInfo, _header_cache -- from
tui/data.py into runs.py, bodies unchanged. The closure was already pure
stdlib + core, and runs.py already imported nearly all of it. tui.data
re-exports every moved name (pinned, so ruff F401 autofix cannot drop them),
so data.<name> stays the TUI-facing spelling and the dashboard is untouched.
cmd_list now calls runs.discover_runs; documents.py types RunInfo from core,
so a non-CLI frontend no longer imports a TUI module to type list_document.

Corrects three false docstrings: tui/__init__.py claimed data was pure-stdlib,
and runs.py pointed twice at ``tui.data.liveness`` for a function now defined
above it.

Guards it with tests/test_entry_point.py::test_extra_less_install_core_works_
and_tui_hints -- a fresh interpreter whose sys.meta_path blocker RAISES for
pyte/rich/textual/tomlkit (returning None would fall through to the dev venv's
real finders and go vacuous, so a vacuity floor pins that), then imports
bmad_loop.cli and runs `list`. Assertions (b)/(d) land in phases 2-3.

Ablation performed: restoring cmd_list's `from .tui.data import discover_runs`
reddens the new test with exactly the reported bug -- child stderr "error: No
module named 'pyte'", rc 1. cli.py restored from a cp backup, diff-verified
byte-identical.

Deviations from the phase plan, all verified against the code:
- data.py keeps its `list_run_dirs` import: the plan's drop-list assumed the
  only user moved, but the missed-decisions cache still calls it (data.py:801).
- Three moved tests referenced data.RunWatcher, not the one the plan
  anticipated; test_classify_crashed and test_classify_legacy_crash_stays_
  interrupted are split the same way, their watcher halves staying behind as
  named companions.
- test_runs.py already had _make_state_run and _dead_pid, semantically
  identical to the helpers the plan said to copy from test_tui_data.py; the
  moved tests reuse them instead of duplicating.
- Removed three imports from test_tui_data.py (pytest, tmux_base, write_pid)
  that went unused when their tests moved.

Verification: full suite 6407 passed / 47 skipped; pyright 0 errors; trunk
check clean. test_tui_app.py run in full, as the plan required: the 35
data.liveness monkeypatch sites still bind through the re-export.
